# src/main/java/com/jeanlucas/mailboxmanager/DTOs/FolderDTO.java
package com.jeanlucas.mailboxmanager.DTOs;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class FolderDTO {

    private int idt;

    private String name;
}