# src/main/java/com/jeanlucas/mailboxmanager/openapi/MailBoxManager-api.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MailBoxManager",
        "description": "API para gestão de caixas de e-mail, pastas e mensagens",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/v1/mailboxes/{mailbox}/folders/{folderIdt}/messages/{messageIdt}/read": {
            "put": {
                "tags": [
                    "message-controller"
                ],
                "operationId": "updateMessageReadStatus",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderIdt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "messageIdt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mailboxes": {
            "get": {
                "tags": [
                    "mail-box-controller"
                ],
                "operationId": "getAllMainBoxes_1",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MailBoxDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "mail-box-controller"
                ],
                "operationId": "createMailBox",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MailBoxDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MailBoxDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mailboxes/{mailbox}/send-message": {
            "post": {
                "tags": [
                    "message-controller"
                ],
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mailboxes/{mailbox}/receive-message": {
            "post": {
                "tags": [
                    "message-controller"
                ],
                "operationId": "saveReceiveMesssage",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mailboxes/{mailbox}/folders": {
            "get": {
                "tags": [
                    "folder-controller"
                ],
                "operationId": "getFoldersByMainBox_1",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FolderDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "folder-controller"
                ],
                "operationId": "createFolder",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FolderDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FolderDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/mailboxes": {
            "get": {
                "tags": [
                    "mail-box-controller"
                ],
                "operationId": "getAllMainBoxes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageMailBoxDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/mailboxes/{mailbox}/folders": {
            "get": {
                "tags": [
                    "folder-controller"
                ],
                "operationId": "getFoldersByMainBox",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageFolderDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/mailboxes/{mailbox}/folders/{folderIdt}/messages": {
            "get": {
                "tags": [
                    "message-controller"
                ],
                "operationId": "getMessagesByMainBoxAndFolder",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderIdt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageMessageDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mailboxes/{mailbox}/folders/{folderIdt}/messages": {
            "get": {
                "tags": [
                    "message-controller"
                ],
                "operationId": "getMessagesByMainBoxAndFolder_1",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderIdt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MessageDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mailboxes/{mailbox}/folders/{folderIdt}/messages/{messageIdt}": {
            "get": {
                "tags": [
                    "message-controller"
                ],
                "operationId": "getMessageDetail",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folderIdt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "messageIdt",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageDTO"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "MessageDTO": {
                "type": "object",
                "properties": {
                    "idt": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sender": {
                        "type": "string"
                    },
                    "recipient": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "read": {
                        "type": "boolean"
                    },
                    "sendAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "folder": {
                        "type": "string"
                    }
                }
            },
            "MailBoxDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    }
                }
            },
            "FolderDTO": {
                "type": "object",
                "properties": {
                    "idt": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "PageMailBoxDTO": {
                "type": "object",
                "properties": {
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MailBoxDTO"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SortObject"
                        }
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SortObject"
                        }
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "direction": {
                        "type": "string"
                    },
                    "nullHandling": {
                        "type": "string"
                    },
                    "ascending": {
                        "type": "boolean"
                    },
                    "property": {
                        "type": "string"
                    },
                    "ignoreCase": {
                        "type": "boolean"
                    }
                }
            },
            "PageFolderDTO": {
                "type": "object",
                "properties": {
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FolderDTO"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SortObject"
                        }
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageMessageDTO": {
                "type": "object",
                "properties": {
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MessageDTO"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SortObject"
                        }
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}